feat: error handling

+++ src/main/java/com/example/suprasysapi/modules/clients/usecases/delete_client/DeleteClientUsecase.java
@@ -1,6 +1,10 @@
 package com.example.suprasysapi.modules.clients.usecases.delete_client;
 
+import java.util.Optional;
+
+import com.example.suprasysapi.modules.clients.entities.Client;
 import com.example.suprasysapi.modules.clients.infra.repositories.ClientRepository;
+import com.example.suprasysapi.shared.exceptions.BadRequestException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,6 +16,12 @@ public class DeleteClientUsecase {
     private ClientRepository repository;
 
     public void execute(Integer id) {
+        Optional<Client> client = repository.findById(id);
+        
+        if (!client.isPresent()) {
+            throw new BadRequestException("Unable to delete the client because it doesn't exist!");
+        }
+
         repository.deleteById(id);
     }
 }
